GH-14: Use the latest official version of grunt-eslint.
Removing the need to use the fluid-grunt-eslint fork. However this does include a fork of the stylish formatter to output the number of linted files when all pass.

In addition to removing the forked grunt task, it allows for tracking newer versions of eslint; which provides access to linting of modern JS syntax.